Add Blink FW to Test All Outputs
Added blink firmware to allow the user to test all outputs of the TTL control module.

The blink FW turns all LED outputs on for a 1/4 second then off for 1 second in a continuous loop.

Beware that any device connected to an HPCM will be on for 1 second and off for a 1/4 second since the HPCM inverts the control logic.